Feat/cleaning

+++ src/app/profile/actions.ts
@@ -3,11 +3,18 @@
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 import prisma from "@/lib/db/prisma";
+import { isReciever } from "@/src/actions/actions";
 
 export const DeleteInvite = async (inviteId: string) => {
+  const invite = await prisma.invite.findUnique({ where: { id: inviteId } });
+
+  if (!invite) return;
+
+  if (!(await isReciever(invite.userEmail))) return;
+
   await prisma.invite.delete({ where: { id: inviteId } });
-  revalidatePath("/");
-  redirect("/");
+  // revalidatePath("/profile");
+  redirect("/profile");
 };
 
 export const AcceptInvite = async (
@@ -17,12 +24,13 @@ export const AcceptInvite = async (
 ) => {
   const user = await prisma.user.findUnique({ where: { email: userEmail } });
 
-  if (user?.email) {
+  if (user?.email && (await isReciever(user.email))) {
     await prisma.race.update({
       where: { id: raceId },
       data: { participants: { push: user.email } },
     });
   }
+
   await DeleteInvite(inviteId);
-  revalidatePath("/");
+  // revalidatePath("/profile");
 };
